:bug: [#7] Ensure the default value of discriminator field
The raw user-supplied data may be missing a discriminator value if a
field default has been set up for the discriminator field. This value
is added to the 'base' serializer data during parent instance
conversion to_internal_value, but the result of that method is not
passed to the context to determine which polymorphic serializer to
resolve.

For this reason, we need to translate back the cleaned data into the
user input.