:sparkles: Allow explicitly marking discriminator value as unmapped
You can now map a discriminator value to None to signal that it's not
mapped to additional fields.

The code was already prepared to deal with None in case a possible
value of the descriminator field is missing from the mapping, so no
runtime changes are needed. The schema generation needed to be updated
to handle this, and the type annotations as well.