:test_tube: Add regression test for schema warning
When generating the component schemas for viewsets that accept
PUT and PATCH requests, drf-spectacular (by default) splits the
schemas for PATCH and PUT/POST schemas. When it does so, it generates
a different name for the PATCH schema, and when a polymorphic
serializer is part of all those (at least two) schemas, then the
shared structure schemas is registered twice, leading to warnings.

This is because there is no existing registry match in the drf-
spectacular code, and it will fully resolve the serializer again.